+++ src/ui/component/AddAddressOptions/style.less
@@ -8,7 +8,7 @@
     font-size: 12px;
     color: @color-comment-1;
   }
-  .field {
+  .address-option-field {
     margin-bottom: 12px;
     font-weight: 500;
     border: 1px solid #ffffff;
@@ -43,7 +43,7 @@
   margin-bottom: 12px;
   background: #ffffff;
   border-radius: 6px;
-  .field {
+  .address-option-field {
     margin-bottom: 12px;
     font-weight: 500;
     border: 1px solid #ffffff;
